- fixed warnings pointed out by Frank
SVN-Revision: 1957

+++ src/level.cpp
@@ -172,8 +172,8 @@ Level::get_total_coins()
 {
   int total_coins = 0;
   for(Sectors::iterator it = sectors.begin(); it != sectors.end(); ++it)
-    for(int x = 0; x < it->second->solids->get_width(); x++)
-      for(int y = 0; y < it->second->solids->get_height(); y++)
+    for(int x = 0; static_cast<unsigned int>(x) < it->second->solids->get_width(); x++)
+      for(int y = 0; static_cast<unsigned int>(y) < it->second->solids->get_height(); y++)
         if(it->second->solids->get_tile(x,y)->attributes & Tile::COIN)
           total_coins++;
   return total_coins;
